datageneration/Instruction/requirements.txt
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
numpy>=1.24.0

